+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/util.dart
@@ -187,6 +187,6 @@ void drawSkShadow(
     devicePixelRatio * ckShadowLightRadius,
     tonalColors.ambient,
     tonalColors.spot,
-    flags,
+    flags.toDouble(),
   );
 }
